# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TecnoCel Web API",
    "description": "API REST para la tienda en línea de productos tecnológicos TecnoCel Web. Incluye gestión de productos, clientes, ventas, carrito de compras y más.",
    "version": "1.0.0",
    "contact": {
      "name": "Soporte TecnoCel"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000/api",
      "description": "Servidor de desarrollo"
    }
  ],
  "securityDefinitions": {
    "bearerAuth": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token. Formato: \"Bearer {token}\""
    },
    "bearerAuthCliente": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header",
      "description": "JWT token para clientes. Formato: \"Bearer {token}\""
    }
  },
  "paths": {
...
    "/ventas/admin/estadisticas": {
      "get": {
        "summary": "Obtener estadísticas de ventas",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Estadísticas de ventas"
          }
        }
      }
    },
    "/ventas/admin/listar": {
      "get": {
        "summary": "Listar ventas",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de ventas"
          }
        }
      }
    },
    "/ventas/admin/registrar": {
      "post": {
        "summary": "Registrar venta manual",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cliente_id": {
                    "type": "integer"
                  },
                  "items": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Venta registrada"
          }
        }
      }
    },
    "/ventas/admin/{id_venta}/cancelar": {
      "patch": {
        "summary": "Cancelar venta",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_venta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Venta cancelada"
          }
        }
      }
    },
    "/ventas/admin/{id_venta}/estado": {
      "patch": {
        "summary": "Actualizar estado de venta",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_venta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "estado": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Estado actualizado"
          }
        }
      }
    },
    "/ventas/admin/tipo-cambio": {
      "get": {
        "summary": "Obtener tipo de cambio",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Tipo de cambio"
          }
        }
      },
      "put": {
        "summary": "Actualizar tipo de cambio",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "cotizacion": {
                    "type": "number"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Tipo de cambio actualizado"
          }
        }
      }
    },
    "/ventas/admin/{id_venta}/comprobante": {
      "get": {
        "summary": "Descargar comprobante",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_venta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "PDF del comprobante"
          }
        }
      }
    },
    "/ventas/admin/{id_venta}/enviar-comprobante": {
      "post": {
        "summary": "Enviar comprobante por email",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_venta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Email enviado"
          }
        }
      }
    },
    "/ventas/admin/{id_venta}": {
      "get": {
        "summary": "Detalle de venta",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_venta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la venta"
          }
        }
      }
    },
    "/ventas/historial": {
      "get": {
        "summary": "Historial de ventas del cliente",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "responses": {
          "200": {
            "description": "Historial de ventas"
          }
        }
      }
    },
    "/ventas/{id_venta}": {
      "get": {
        "summary": "Detalle de venta de cliente",
        "tags": [
          "Ventas"
        ],
        "security": [
          {
            "bearerAuthCliente": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id_venta",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Detalle de la venta"
          }
        }
      }
    }
  },
...
}